+++ frontend/src/components/layout/MainNav/MainNav.tsx
@@ -2,16 +2,21 @@ import React from "react";
 import { useTranslation } from "react-i18next";
 import { NavLink } from "react-router-dom";
 
-import { NAV_ITEMS } from "constants/navigation";
+import { GUEST_NAV_ITEMS, NAV_ITEMS } from "constants/navigation";
+
+import { useAppSelector } from "redux/hooks";
+import { selectIsGuest } from "redux/selectors/viewerSelectors";
 
 import styles from "./MainNav.module.scss";
 
 export const MainNav: React.FC = () => {
     const { t } = useTranslation();
+    const isGuest = useAppSelector(selectIsGuest);
+    const items = isGuest ? GUEST_NAV_ITEMS : NAV_ITEMS;
 
     return (
         <nav className={styles["main-nav"]}>
-            {NAV_ITEMS.map(({ to, labelKey }) => (
+            {items.map(({ to, labelKey }) => (
                 <NavLink
                     key={to}
                     to={to}